Relax ssh config file parsing
This commit relaxes the handling of ssh config files. Rex used to fail
to parse my ssh config due to my habit of placing equal signs
between the keys and the values (which is perfectly fine with ssh).

It also introduces trimming of values read from this file and some testcases
covering this patch.