Support password-less sudo
This commit implements a workaround to make Rex work
with limit sudo rights without a password.

There are setups where certain users are allowed to
use only some commands via sudo but those only without
a password. Those users may even not have any password,
so sudo with a password will obviously not work.

The drawback of this commit is that it looses some features
that come with the former approach (like setting LC_ALL and PATH).