Optimise Rex::Commands::Augeas
In the case of SSH, this patch groups Augeas commands and runs
them together. Starting augtool is expensive, which is now done
only once for each group of commands.

In addition, the commands are passed via a file, fixing bugs
if the commands have characters that would otherwise need
to be escaped.