Ensure update_system always updates package repositories first
There is a mixture of behaviours with update_system. With some
operating systems it will update the package repository cache
first, with others it won't. This patch attempts to make the
behaviour consistent, and always update repositories first, given
that this is most likely what one would want to do when updating
a system.

Disclaimer: I have only experience with apt, but have attempted
to cover all operating systems.