Don't recreate connection when executing rethink_connection.
We need to create a task object's connection as late as possible, so
that everything is setup that it relies on. When modifying an of a tasks
hooks, it would run rethink_connection which in turn would re-create a
local connection before it knew it was supposed to be an ssh connection.