Don't recreate connection during rethink_connection (fix #694, close #727)
We need to create a task object's connection as late as possible, so
that everything is setup that it relies on. When modifying any of a task's
hooks, it would run rethink_connection which in turn would recreate a
local connection before it knew it was supposed to be an SSH connection.