A way for adding several ssh keys, also from file.
This patch makes it possible to add several keys to the users
authorized_keys and also makes it possible to either append or
overwrite.

This is very useful when several people are using Rex for handling
the same systems.

An example of how it could be used:

task "my_user", sub {

  my $user = getlogin || getpwuid($<);
  my @key_files = (
    "/root/.ssh/id_rsa.pub",
    "/home/$user/.ssh/id_rsa.pub",
  );

  my @ssh_keys;
  for my $file (@key_files) {
    open my $fh, "<", "$file";
    while (my $line = <$fh>) {
      chomp $line;
      push(@ssh_keys, "$line");
    }
  }

  create_user "server_user",
              home            => '/home/server_user',
              comment         => 'Demo user',
              expire          => '2100-05-30',
              password        => 'password',
              system          => 1,
              create_home     => TRUE,
              ssh_keys_append => TRUE,
              ssh_keys        => \@ssh_keys;
};