Detect services that are not already on runlevels
Services that are not enabled in a runlevel are terminated with a
newline in `rc-config list`. This newline does not match the grep
pattern "\s" and Rex reports that the service does not exist.

`rc-config list` contains a header that needs to be ignored, then
grab the first space separated word and match the provided service
name.

	$ rc-config list | grep sshd
	  sshd
	$ rc-config list | grep "\ssshd\s"
	$ rc-config list | awk '!/^(Available i|I)/ && $1 ~ /^sshd$/{print$1}'
	sshd
	$ rc-config list | grep "\sinit\s"
	Available init scripts
	$ rc-config list | awk '!/^(Available i|I)/ && $1 ~ /^init$/{print$1}'
	$