Added Rex::RunList.
A Rex::TaskList object is a list of all possible tasks defined by the Rexfile.
A Rex::RunList object is a list of all the tasks that will get executed during
this run.  In the past there was only one instance of any given task.  The
RunList concetp allows users to run a task more than once in a single run.  It
also adds the concepts of opts and args to the Rex::Task object.