Fix Pkgconf for Debian
For reasons I do not understand, PkgConf for Debian adds 2 lines for
each question, which causes errors with debconf-set-selections and is
not as per the manual. It was me that wrote that line in originally, and
I don't know why I did it or why it didn't fail before.

Anyway, this patch removes it, and allows it to function correctly.