Add ability to pass options to openssh via REX_OPENSSH_OPTS
As an example, per RFC, IPv6 takes precedent over IPv4, but is sometimes broken
in various exciting ways. In that case it would be useful to pass -4 to ssh.

Since all breakage points are hard to predict, it'd make sense to do it like
git, e.g. set REX_OPENSSH_OPTS and avoid further opt parsing.