Add tools in contrib/ to strip dist.ini in-place for Gentoo
The implementation of this is slightly more complicated than strictly
necessary, but it adds a lot of value in output transparency and the
ability to error when things that were expected to be removed, failed
to remove, for whatever reason.

The idea is that in gentoo's live ebuilds, instead of this logic
being embedded in the ebuild, we can invoke this during src_prepare,
and subsequently, give the git repository far more control of keeping
this problem under wraps, as when plugins are added that pose a
problem, they can be removed again by this script in order to get the
gentoo live ebuilds working again for your users, without needing to
push more changes through gentoo.

I've tried to make the various extraction tools as vendor agnostic
as possible, and that way, if need be, other vendors can have their own
script in this dir which nukes different things.

Also, these contributed files are currently configured in [GatherDir]
such that they are trimmed from the published release, as they really
don't have much value there, because as it is, many projects will *not*
work when running "dzil build" on the published sources, as vast
quantities of assumptions it makes about "things not being changed yet"
fail to be true.