Remove unused helper script
This helper script was introduced on
8b499f5f0ac29f81f5912562056410090b7f79e4, went through several
modifications, and was retired 5 month later on
f13b970cfa385f5993ef458e6513a535ccad065a "because it doesn't work as
expected".

It is getting removed now due to being considered an internal
implementation detail which is not used by the code base since several
years.

In case your code relies on it, please open a bug.