This PR is a proposal to fix #1609
As per #1609 it is not possible to retrieve errors from Augeas if it is
failing. This PR adds a new command to be able to retrieve them.

I have not added a test, as there are currently no tests at all for the
Augeas command.

I thought about automatically retrieving and adding any error
information to a failing Augeas command, but I decided that is
over-engineering things too much and that a developer should decide how
they want to handle errors and retrieve them manually if they wish to.

Please review and merge, or let me know how to improve it further.