Allow libxml2 version warnings for CPAN Testers
According to its documentation, “XML::LibXML issues a warning if the
version of libxml2 dynamically linked to it is less than the version of
libxml2 which it was compiled against.”

In general, we would like to fail the test suite on such warnings to
require a conscious decision about how to resolve the situation.
However, on automated testing systems, like CPAN Testers, we have
limited control over applying fixes.

Accepting this warning on such systems reduces the amount of failing
reports we can’t fix, while still failing elsewhere as intended.

When encountering this issue on an automated testing system, please
consider notifying its owner about it.